Add LinSpace Operator
Summary:
This commit adds the linspace op for dragon, torch and tensorflow.
And, a workaround for truncated int interval is made to range/linspace (up to 2**57).

+++ dragon/kernels/array/range_op_kernel.cu
@@ -11,20 +11,20 @@ namespace {
 
 template <typename T>
 __global__ void
-_Range(const int nthreads, const float start, const float delta, T* y) {
+_Range(const int nthreads, const double start, const double delta, T* y) {
   CUDA_1D_KERNEL_LOOP(i, nthreads) {
-    y[i] = T(start + (float)i * delta);
+    y[i] = T(start + double(i) * delta);
   }
 }
 
 template <>
 __global__ void _Range<half>(
     const int nthreads,
-    const float start,
-    const float delta,
+    const double start,
+    const double delta,
     half* y) {
   CUDA_1D_KERNEL_LOOP(i, nthreads) {
-    y[i] = __float2half(start + (float)i * delta);
+    y[i] = __float2half(float(start + double(i) * delta));
   }
 }
 
@@ -35,8 +35,8 @@ __global__ void _Range<half>(
 template <>
 void Range<float16, CUDAContext>(
     const int count,
-    const float start,
-    const float delta,
+    const double start,
+    const double delta,
     float16* y,
     CUDAContext* ctx) {
   _Range<<<CUDA_BLOCKS(count), CUDA_THREADS, 0, ctx->cuda_stream()>>>(
@@ -47,8 +47,8 @@ void Range<float16, CUDAContext>(
   template <>                                                            \
   void Range<T, CUDAContext>(                                            \
       const int count,                                                   \
-      const float start,                                                 \
-      const float delta,                                                 \
+      const double start,                                                \
+      const double delta,                                                \
       T* y,                                                              \
       CUDAContext* ctx) {                                                \
     _Range<<<CUDA_BLOCKS(count), CUDA_THREADS, 0, ctx->cuda_stream()>>>( \
